refactor (gravsearch): move global methods from SearchResponderV2 to GravsearchUtilV2
